requirements.txt
sentence-transformers
faiss-cpu
transformers
numpy
datasets
peft
accelerate
scipy
torch
